re #15 - Set the default layout and model depending on how com_logs is called.
Basically, com_logs can be called as <?=KFactory::tmp('admin::com.logs.controller.log')->package('componentname')->display();?> in a template. This allows com_logs to be embedded inside components. If it was called through HMVC, it will use a special layout.

The controller also makes sure that the com_logs model is used.