Code cleanup.
Added user input checks. A list of ids can be passed so that words from those inputs are included in the dictionary. This is useful for penalizing passwords containing these words. However the dictionary seems to behave in a persistent way, i.e. words added are kept for future checks. This sounds like a bug on zxcvbn lib as this makes no sense. Disabling behavior for now.

re #96.