Refactored command package. Renamed 'handler' to 'callback' and 'invoker' to 'handler.
A command handler is a mixin and can delegate execution of a callback to an object if that object implements the CommandCallbackDelegate interface.

The command mixin itself is now a command handler, when executed it will invoke the callbacks that are registered through it.

The command chain no longer returns an array. If a break condition is reached the condition will be returned, otherwise the chain will return NULL.